copilot review changes - separate composable for array sorting
